Fix Flux2Klein export: BN stats float conversion and text encoder forward bypass
- Convert BN running_mean/running_var to float32 before numpy (bfloat16 not supported)
- Bypass Qwen3 create_causal_mask in text encoder forward (incompatible with torch.jit.trace)
- Directly iterate through model layers collecting hidden states from target layers

+++ optimum/exporters/openvino/convert.py
@@ -770,8 +770,8 @@ def export_from_model(
 
             np.savez(
                 output / "vae_bn_stats.npz",
-                running_mean=vae.bn.running_mean.cpu().numpy(),
-                running_var=vae.bn.running_var.cpu().numpy(),
+                running_mean=vae.bn.running_mean.cpu().float().numpy(),
+                running_var=vae.bn.running_var.cpu().float().numpy(),
             )
 
     _set_runtime_options(
